requirements.txt
fastapi
python-socketio
uvicorn
websockets
motor
random-object-id
pymongo[srv]
gunicorn
jinja2